Move all the copy/paste token selection stuff into new class.
